# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 电力交易智能辅助决策系统 API
  description: |
    配套开发基线 5.5 API 目录总表。统一响应封装 ApiResponse；分页用 PageResult；
    鉴权 JWT Bearer；错误码分段见 README 第 4 条。所有路径省略网关前缀 /api/v1。
  version: 1.0.0
  license:
    name: 项目内部使用
    url: https://api.ptidss.example.com/license
  contact:
    name: PTIDSS 架构组
servers:
  - url: https://api.ptidss.example.com/api/v1
    description: 生产（示例域名）
  - url: https://api-test.ptidss.example.com/api/v1
    description: 测试
...
security:
  - bearerAuth: []
paths:
...
  /intel/news:
    get:
      operationId: getIntelNews
      summary: 情报流（归一化标签/重要度筛选）
      tags: [intel]
      parameters:
        - { name: importance, in: query, schema: { type: string, enum: [high, medium, low] } }
        - { name: intelType, in: query, schema: { type: string, enum: [price, weather, supply_demand, policy, announcement, opinion] } }
        - { $ref: '#/components/parameters/PageParam' }
      responses:
        '200':
          description: 情报列表
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties: { data: { $ref: '#/components/schemas/PageResult' } }
        '4XX': { $ref: '#/components/responses/BizError' }
  /intel/sources:
    get:
      operationId: getIntelSources
      summary: 情报源配置与状态
      tags: [intel]
      responses:
        '200':
          description: 情报源
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data:
                        type: array
                        items: { type: object }
        '4XX': { $ref: '#/components/responses/BizError' }
  /intel/push-rules:
    get:
      operationId: getIntelPushRules
      summary: 推送规则列表
      tags: [intel]
      responses:
        '200':
          description: 规则列表
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data:
                        type: array
                        items: { type: object }
        '4XX': { $ref: '#/components/responses/BizError' }
    post:
      operationId: postIntelPushRules
      summary: 配置情报推送规则（标签+重要度→渠道/角色）
      tags: [intel]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ruleName, matchTags, importance]
              properties:
                ruleName: { type: string }
                matchTags: { type: array, items: { type: string } }
                importance: { type: string, enum: [high, medium, low] }
                targets: { type: array, items: { type: string }, description: 角色/用户编码 }
      responses:
        '200':
          description: 规则已配置
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data:
                        type: object
                        properties: { ruleId: { type: string } }
        '4XX': { $ref: '#/components/responses/BizError' }
  /intel/push-rules/execute:
    post:
      operationId: postIntelPushRulesExecute
      summary: 手动触发推送规则执行（情报→消息中心联动）
      description: >
        执行全部 active 推送规则：匹配未推送情报（标签交集 × 重要度一致）→ 按目标角色
        派发个人消息（msg_type=intel_push）→ 更新情报推送状态。另有服务端定时 30s 自动执行兜底。
      tags: [intel]
      responses:
        '200':
          description: 推送执行完成
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data:
                        type: object
                        properties:
                          matchedNews: { type: integer, description: 匹配并推送的情报条数 }
                          pushedMessages: { type: integer, description: 派发的个人消息条数 }
        '4XX': { $ref: '#/components/responses/BizError' }

  # ==================== agent（智能体编排，V2.1） ====================
  /agent/registry:
    get:
      operationId: getAgentRegistry
      summary: 智能体注册表（7 大智能体：态势/预测/报价/合规/风险/收益/复盘）
      tags: [agent]
      responses:
        '200':
          description: 注册表
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data:
                        type: array
                        items: { type: object, description: agentCode/名称/模型绑定/状态/置信度 }
        '4XX': { $ref: '#/components/responses/BizError' }
  /agent/runs:
    get:
      operationId: getAgentRuns
      summary: 智能体运行记录（输入快照/输出/置信度/耗时/状态）
      tags: [agent]
      responses:
        '200':
          description: 运行记录
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/AgentRun' }
        '4XX': { $ref: '#/components/responses/BizError' }
...
  /agent/registry/{id}/status:
    post:
      operationId: postAgentStatus
      summary: 智能体启停（仅 admin 角色，V3G-01 收紧）
      tags: [agent]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { status: { type: string, enum: [active, disabled] } }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
  /agent/registry/{id}/model-config:
    post:
      operationId: postAgentModelConfig
      summary: 智能体模型/LLM 绑定（仅 admin 角色，V3G-02 收紧）
      tags: [agent]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { modelCode: { type: string } }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }

  # ==================== trade 网关（V2.2，密钥类敏感配置） ====================
...
  /admin/configs:
    get:
      operationId: getAdminConfigs
      summary: 系统配置列表（分组/关键字搜索；敏感项脱敏 ******）
      tags: [admin]
      parameters:
        - { name: group, in: query, schema: { type: string, description: 配置组编码 } }
        - { name: keyword, in: query, schema: { type: string, description: 名称/键模糊搜索 } }
      responses:
        '200':
          description: 配置列表
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/SysConfig' }
        '4XX': { $ref: '#/components/responses/BizError' }
    post:
      operationId: postAdminConfigs
      summary: 新建系统配置（仅 admin 角色，V3G-09 收紧）
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SysConfig' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
  /admin/configs/{id}:
    get:
      operationId: getAdminConfigsDetail
      summary: 配置详情（敏感项脱敏）
      tags: [admin]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: 详情
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties: { data: { $ref: '#/components/schemas/SysConfig' } }
        '4XX': { $ref: '#/components/responses/BizError' }
    put:
      operationId: putAdminConfigsDetail
      summary: 更新配置值（内置项禁改 key 可改值/启停；仅 admin 角色，V3G-09 收紧）
      tags: [admin]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SysConfig' }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }
    delete:
      operationId: deleteAdminConfigsDetail
      summary: 删除配置（内置项禁删；仅 admin 角色，V3G-09 收紧）
      tags: [admin]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiResponse' }
        '4XX': { $ref: '#/components/responses/BizError' }

  # ==================== intel 采集执行/状态（V2.5 后端就绪，V3.2 前端入口补全） ====================
  /intel/fetch-status:
    get:
      operationId: getIntelFetchStatus
      summary: 行情采集状态监测（各源最近成功/失败原因/连续失败/频率/端点域名）
      tags: [intel]
      responses:
        '200':
          description: 采集状态列表
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/IntelFetchStatus' }
        '4XX': { $ref: '#/components/responses/BizError' }
  /intel/fetch:
    post:
      operationId: postIntelFetch
      summary: 手动触发行情采集（force=true 忽略频率立即全量重跑；仅 admin 角色）
      tags: [intel]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties: { force: { type: boolean, description: true=忽略频率强制全量 } }
      responses:
        '200':
          description: 采集结果
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - properties: { data: { type: object } }
        '4XX': { $ref: '#/components/responses/BizError' }

  # ==================== flow 流程引擎（V2.2 定义化） ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: pageNo
      in: query
      description: 分页参数（pageNo/pageSize，对齐后端 @RequestParam 实现）
      schema:
        type: object
        properties:
          pageNo: { type: integer, minimum: 1, default: 1 }
          pageSize: { type: integer, maximum: 200, default: 20 }
  responses:
    BizError:
      description: 业务错误（错误码分段见 README）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message, data, traceId]
      properties:
        code: { type: integer, description: 0 成功；非 0 见错误码分段表, example: 0 }
        message: { type: string, example: success }
        data: { type: object, nullable: true, description: 业务数据（随接口变化） }
        traceId: { type: string, description: 全链路追踪 ID }
    ErrorResponse:
      type: object
      properties:
        code: { type: integer, example: 14001 }
        message: { type: string }
        traceId: { type: string }
        detail: { type: object, nullable: true, description: 参数错误明细 }
    PageResult:
      type: object
      properties:
        list: { type: array, items: { type: object } }
        pageNo: { type: integer }
        pageSize: { type: integer }
        total: { type: integer }
...
    AgentRun:
      type: object
      properties:
        agentCode: { type: string }
        inputSnapshot: { type: object }
        output: { type: object }
        confidence: { type: number }
        elapsedMs: { type: integer }
        status: { type: string }
...
    SysConfig:
      type: object
      description: 系统配置中心条目（V3.0，33 项内置种子；敏感项值加密存储/脱敏展示）
      properties:
        id: { type: integer }
        groupCode: { type: string, description: 配置组：rule/settlement/region/optimize/forecast/model/agent/security/notice }
        configKey: { type: string, description: 配置键（内置项禁改） }
        configName: { type: string }
        valueType: { type: string, enum: [string, number, boolean, select, json] }
        value: { type: string, description: 敏感项返回 ****** }
        defaultValue: { type: string }
        options: { type: string, description: select 枚举范围，逗号分隔 }
        enabled: { type: boolean }
        builtIn: { type: boolean, description: true=内置项禁删禁改 key }
        remark: { type: string }
        updatedAt: { type: string }
    IntelFetchStatus:
      type: object
      description: 行情采集状态监测项（V2.5）
      properties:
        id: { type: string }
        sourceCode: { type: string }
        sourceName: { type: string }
        intelType: { type: string }
        fetchMode: { type: string, enum: [api, crawl, file] }
        frequency: { type: string }
        status: { type: string, enum: [enabled, disabled] }
        lastSuccessAt: { type: string, nullable: true }
        lastError: { type: string, nullable: true }
        consecutiveFailures: { type: integer }
        frequencyMinutes: { type: string }
        endpoint: { type: string, description: 端点域名（脱敏主机） }
        mock: { type: boolean, description: true=模拟采集路径 }
        healthy: { type: boolean, description: 启用且连续失败 < 阈值 }